# docs/07_interfaces/openapi_v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DevPilot Local API Contract v1",
    "version": "1.0.0-implemented-initial",
    "description": "Static OpenAPI contract synchronized with the implemented-initial local FastAPI MVP. FUNC-SPRINT-67 implements the server adapter on localhost; Sprint 68 will add token/CORS hardening."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787",
      "description": "Future local-only server. Not active in Sprint 66."
    }
  ],
...
  "paths": {
    "/api/v1/workspace/status": {
      "get": {
        "operationId": "workspace_status",
        "summary": "API-WORKSPACE-STATUS: workspace.status",
        "description": "Contract-only endpoint mapped to `WorkspaceApplicationService.status`. Sprint 67 implements this route in the local FastAPI MVP.",
        "tags": [
          "workspace"
        ],
        "x-devpilot-api-id": "API-WORKSPACE-STATUS",
        "x-devpilot-operation": "workspace.status",
        "x-devpilot-domain-service": "WorkspaceApplicationService.status",
        "x-devpilot-status": "implemented-initial",
        "x-devpilot-side-effect": "none",
        "x-devpilot-dry-run-default": true,
        "x-devpilot-auth": "public-local-read",
        "x-devpilot-policy": "None",
        "x-devpilot-command-equivalent": "python -m devpilot_core workspace status --json",
        "responses": {
          "200": {
            "description": "ApplicationResponse wrapping CommandResult.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request envelope or unsupported parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "403": {
            "description": "Policy/approval/security gate blocked the operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Contract validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "500": {
            "description": "Controlled internal error preserving response envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          }
        },
        "parameters": []
      }
    },
...
    "/api/v1/miasi/status": {
      "get": {
        "operationId": "miasi_validate",
        "summary": "API-MIASI-STATUS: miasi.validate",
        "description": "Contract-only endpoint mapped to `MiasiApplicationService.validate`. Sprint 67 implements this route in the local FastAPI MVP.",
        "tags": [
          "miasi"
        ],
        "x-devpilot-api-id": "API-MIASI-STATUS",
        "x-devpilot-operation": "miasi.validate",
        "x-devpilot-domain-service": "MiasiApplicationService.validate",
        "x-devpilot-status": "implemented-initial",
        "x-devpilot-side-effect": "none",
        "x-devpilot-dry-run-default": true,
        "x-devpilot-auth": "public-local-read",
        "x-devpilot-policy": "MIASI registries read-only",
        "x-devpilot-command-equivalent": "python -m devpilot_core miasi validate --json",
        "responses": {
          "200": {
            "description": "ApplicationResponse wrapping CommandResult.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request envelope or unsupported parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "403": {
            "description": "Policy/approval/security gate blocked the operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Contract validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "500": {
            "description": "Controlled internal error preserving response envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          }
        },
        "parameters": []
      }
    },
    "/api/v1/repo/inventory": {
      "get": {
        "operationId": "repo_inventory",
        "summary": "API-REPO-INVENTORY: repo.inventory",
        "description": "Contract-only endpoint mapped to `RepoApplicationService.inventory`. Sprint 67 implements this route in the local FastAPI MVP.",
        "tags": [
          "repo"
        ],
        "x-devpilot-api-id": "API-REPO-INVENTORY",
        "x-devpilot-operation": "repo.inventory",
        "x-devpilot-domain-service": "RepoApplicationService.inventory",
        "x-devpilot-status": "implemented-initial",
        "x-devpilot-side-effect": "none",
        "x-devpilot-dry-run-default": true,
        "x-devpilot-auth": "token-planned-sprint-68",
        "x-devpilot-policy": "Git read-only / PathGuard",
        "x-devpilot-command-equivalent": "python -m devpilot_core repo inventory --json",
        "responses": {
          "200": {
            "description": "ApplicationResponse wrapping CommandResult.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request envelope or unsupported parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "403": {
            "description": "Policy/approval/security gate blocked the operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Contract validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "500": {
            "description": "Controlled internal error preserving response envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          }
        },
        "parameters": []
      }
    },
...
    "/api/v1/model/providers": {
      "get": {
        "operationId": "model_providers",
        "summary": "API-MODEL-PROVIDERS: model.providers",
        "description": "Contract-only endpoint mapped to `ModelApplicationService.providers`. Sprint 67 implements this route in the local FastAPI MVP.",
        "tags": [
          "model"
        ],
        "x-devpilot-api-id": "API-MODEL-PROVIDERS",
        "x-devpilot-operation": "model.providers",
        "x-devpilot-domain-service": "ModelApplicationService.providers",
        "x-devpilot-status": "implemented-initial",
        "x-devpilot-side-effect": "none",
        "x-devpilot-dry-run-default": true,
        "x-devpilot-auth": "token-planned-sprint-68",
        "x-devpilot-policy": "CostGuard/SecretGuard; no external API",
        "x-devpilot-command-equivalent": "python -m devpilot_core model providers --json",
        "responses": {
          "200": {
            "description": "ApplicationResponse wrapping CommandResult.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request envelope or unsupported parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "403": {
            "description": "Policy/approval/security gate blocked the operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Contract validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "500": {
            "description": "Controlled internal error preserving response envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          }
        },
        "parameters": []
      }
    },
    "/api/v1/observability/traces": {
      "get": {
        "operationId": "observability_trace_report",
        "summary": "API-OBSERVABILITY-TRACES: observability.trace_report",
        "description": "Contract-only endpoint mapped to `ObservabilityApplicationService.trace_report`. Sprint 67 implements this route in the local FastAPI MVP.",
        "tags": [
          "observability"
        ],
        "x-devpilot-api-id": "API-OBSERVABILITY-TRACES",
        "x-devpilot-operation": "observability.trace_report",
        "x-devpilot-domain-service": "ObservabilityApplicationService.trace_report",
        "x-devpilot-status": "implemented-initial",
        "x-devpilot-side-effect": "none",
        "x-devpilot-dry-run-default": true,
        "x-devpilot-auth": "token-planned-sprint-68",
        "x-devpilot-policy": "bounded read; secret redaction",
        "x-devpilot-command-equivalent": "python -m devpilot_core trace report --json",
        "responses": {
          "200": {
            "description": "ApplicationResponse wrapping CommandResult.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request envelope or unsupported parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "403": {
            "description": "Policy/approval/security gate blocked the operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Contract validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "500": {
            "description": "Controlled internal error preserving response envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            },
            "description": "Optional bounded read limit for endpoints that support it."
          }
        ]
      }
    },
...
    "/api/v1/history/runs": {
      "get": {
        "operationId": "history_runs",
        "summary": "API-HISTORY-RUNS: history.runs",
        "description": "Contract-only endpoint mapped to `HistoryApplicationService.list_runs`. Sprint 67 implements this route in the local FastAPI MVP.",
        "tags": [
          "history"
        ],
        "x-devpilot-api-id": "API-HISTORY-RUNS",
        "x-devpilot-operation": "history.runs",
        "x-devpilot-domain-service": "HistoryApplicationService.list_runs",
        "x-devpilot-status": "implemented-initial",
        "x-devpilot-side-effect": "none",
        "x-devpilot-dry-run-default": true,
        "x-devpilot-auth": "token-planned-sprint-68",
        "x-devpilot-policy": "bounded LocalStore read",
        "x-devpilot-command-equivalent": "python -m devpilot_core history list --json",
        "responses": {
          "200": {
            "description": "ApplicationResponse wrapping CommandResult.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request envelope or unsupported parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "403": {
            "description": "Policy/approval/security gate blocked the operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Contract validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "500": {
            "description": "Controlled internal error preserving response envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            },
            "description": "Optional bounded read limit for endpoints that support it."
          }
        ]
      }
    },
    "/api/v1/application/contract": {
      "get": {
        "operationId": "app_contract",
        "summary": "API-APPLICATION-CONTRACT: app.contract",
        "description": "Contract-only endpoint mapped to `ApplicationService.application_contract`. Sprint 67 implements this route in the local FastAPI MVP.",
        "tags": [
          "application"
        ],
        "x-devpilot-api-id": "API-APPLICATION-CONTRACT",
        "x-devpilot-operation": "app.contract",
        "x-devpilot-domain-service": "ApplicationService.application_contract",
        "x-devpilot-status": "implemented-initial",
        "x-devpilot-side-effect": "none",
        "x-devpilot-dry-run-default": true,
        "x-devpilot-auth": "public-local-read",
        "x-devpilot-policy": "None",
        "x-devpilot-command-equivalent": "python -m devpilot_core app contract --json",
        "responses": {
          "200": {
            "description": "ApplicationResponse wrapping CommandResult.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request envelope or unsupported parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "403": {
            "description": "Policy/approval/security gate blocked the operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Contract validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "500": {
            "description": "Controlled internal error preserving response envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          }
        },
        "parameters": []
      }
    },
    "/api/v1/standards/status": {
      "get": {
        "operationId": "standards_status",
        "summary": "API-STANDARDS-STATUS: standards.status",
        "description": "Implemented-initial endpoint mapped to `ValidationApplicationService.standards_status` through ApplicationService.",
        "tags": [
          "standards"
        ],
        "x-devpilot-api-id": "API-STANDARDS-STATUS",
        "x-devpilot-operation": "standards.status",
        "x-devpilot-domain-service": "ValidationApplicationService.standards_status",
        "x-devpilot-status": "implemented-initial",
        "x-devpilot-side-effect": "none",
        "x-devpilot-dry-run-default": true,
        "x-devpilot-auth": "public-local-read",
        "x-devpilot-policy": "DOC_VALIDATE_ALLOW",
        "x-devpilot-command-equivalent": "python -m devpilot_core standards status --json",
        "responses": {
          "200": {
            "description": "ApplicationResponse wrapping CommandResult.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request envelope or unsupported parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "403": {
            "description": "Policy/approval/security gate blocked the operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Contract validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "500": {
            "description": "Controlled internal error preserving response envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          }
        },
        "parameters": []
      }
    }
  },
  "components": {
    "schemas": {
      "Finding": {
        "type": "object",
        "required": [
          "id",
          "message",
          "severity"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "severity": {
            "type": "string",
            "enum": [
              "info",
              "warning",
              "fail",
              "block",
              "error"
            ]
          },
          "path": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "additionalProperties": true
      },
...
      "ApplicationResponse": {
        "type": "object",
        "required": [
          "contract",
          "schema_version",
          "operation",
          "ok",
          "exit_code",
          "message",
          "data",
          "findings",
          "generated_at"
        ],
        "properties": {
          "contract": {
            "type": "string",
            "const": "DevPilotApplicationResponse"
          },
          "schema_version": {
            "type": "string"
          },
          "operation": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          },
          "exit_code": {
            "type": "integer",
            "enum": [
              0,
              1,
              2,
              3
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          },
          "findings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Finding"
            }
          },
          "generated_at": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
      "ErrorApplicationResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApplicationResponse"
          }
        ],
        "description": "Error envelope preserving ApplicationResponse semantics. Non-2xx responses must still carry operation, exit_code and findings."
      }
    },
    "securitySchemes": {
      "LocalTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-DevPilot-Local-Token",
        "description": "Planned for FUNC-SPRINT-68. Not enforced by Sprint 66 because no HTTP server exists."
      }
    }
  }
}